borrow_or_share

Trait Bos

pub trait Bos<T: ?Sized> {
    type Ref<'this>: Ref<T>
       where Self: 'this;

    // Required method
    fn borrow_or_share(this: &Self) -> Self::Ref<'_>;
}

A trait for either borrowing or sharing data.

See the crate-level documentation for more details.

Required Associated Types

type Ref<'this>: Ref<T> where Self: 'this

The resulting reference type. May only be &T.

Required Methods

fn borrow_or_share(this: &Self) -> Self::Ref<'_>

Borrows from *this or from behind a reference it holds, returning a reference of type Self::Ref.

In the latter case, the returned reference is said to be shared with *this.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.