Crate borrow_or_share

Traits for either borrowing data or sharing references.

Walkthrough

Suppose that you have a struct Text<T> where T may be String or &str. You want to implement a generic method as_str on Text which returns a longest-living reference to the inner string. This is when BorrowOrShare comes in handy:

use borrow_or_share::BorrowOrShare;

struct Text<T>(T);

impl<'i, 'o, T: BorrowOrShare<'i, 'o, str>> Text<T> {
    fn as_str(&'i self) -> &'o str {
        self.0.borrow_or_share()
    }
}

// The returned reference is borrowed from `*t` and lives as long as `t`.
fn owned_as_str(t: &Text<String>) -> &str {
    t.as_str()
}

// The returned reference is copied from `t.0` and lives longer than `t`.
fn borrowed_as_str(t: Text<&str>) -> &str {
    t.as_str()
}

The BorrowOrShare trait takes two lifetime parameters 'i, 'o, and a type parameter T. Its borrow_or_share method takes &'i self and returns &'o T. You can use the trait to write your own “data-borrowing or reference-sharing” functions, like the as_str method in the above example.

The lifetime parameters on BorrowOrShare can be quite restrictive when reference-sharing behavior is not needed, such as in an AsRef implementation. In such cases, Bos should be used instead:

use borrow_or_share::{BorrowOrShare, Bos};

struct Text<T>(T);

impl<'i, 'o, T: BorrowOrShare<'i, 'o, str>> Text<T> {
    fn as_str(&'i self) -> &'o str {
        self.0.borrow_or_share()
    }
}

impl<T: Bos<str>> AsRef<str> for Text<T> {
    fn as_ref(&self) -> &str {
        self.as_str()
    }
}

In the above example, the as_str method is also available on Text<T> where T: Bos<str>, because BorrowOrShare is implemented for all types that implement Bos. It also works the other way round because Bos is a supertrait of BorrowOrShare.

Note that Bos<T> is also implemented for other types such as T, &mut T, Box<T>, and Rc<T>. Consider adding extra trait bounds, preferably on a function that constructs your type, if this is not desirable.

You can also implement Bos for your own type, for example:

use borrow_or_share::Bos;

struct Text<'a>(&'a str);

impl<'a> Bos<str> for Text<'a> {
    type Ref<'this> = &'a str where Self: 'this;
     
    fn borrow_or_share(this: &Self) -> Self::Ref<'_> {
        this.0
    }
}

Relation between Bos, Borrow, and AsRef

Bos is similar to Borrow and AsRef, but different in a few aspects:

The implementation of Bos for &T copies the reference with lifetime unchanged instead of borrowing from it.

Bos does not have extra requirements on Eq, Ord, and Hash implementations as Borrow does.

Despite being safe to implement, Bos is not meant to be eagerly implemented as AsRef is. This crate only provides implementations of Bos for types that currently implement Borrow in the standard library. If this is too restrictive, feel free to copy the code pattern from this crate as you wish.

Crate features

std (default): Enables Bos implementations for OsString and PathBuf.

Traits

BorrowOrShare

A helper trait for writing “data-borrowing or reference-sharing” functions.

Bos

A trait for either borrowing data or sharing references.