Trait borrow_or_share::Bos

pub trait Bos<T: ?Sized> {
    type Ref<'this>: Ref<T>
       where Self: 'this;

    // Required method
    fn borrow_or_share(this: &Self) -> Self::Ref<'_>;
}

A trait for either borrowing data or sharing references.

See the crate-level documentation for more details.

Required Associated Types

type Ref<'this>: Ref<T> where Self: 'this

The resulting reference type. May only be &T.

Required Methods

fn borrow_or_share(this: &Self) -> Self::Ref<'_>

Borrows from a value or gets a shared reference from it, returning a reference of type Self::Ref.

Object Safety

This trait is not object safe.